+++ src/entity/index.ts
@@ -1,2 +1,3 @@
 export * from "./icon.js";
 export * from "./list.js";
+export * from "./registry.js";